Add cross-test for Silero VAD pretrained model and unify reference model methods in `bunsen-silero-onnx`.
